+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `xan separate` will not trim splitted values with some modes by default anymore.
 * Dropping `xan network --stats` in favor of `-f stats`.
 * `-D` becomes short flag for `xan network --degrees` instead of `--disjoint-keys`.
+* `xan separate --capture-groups` is dropped in favor of `-c/--captures` & `-C/--all-captures`.
 
 *Features*